requirements.txt
appnope==0.1.3
asgiref==3.6.0
asttokens==2.2.1
backcall==0.2.0
backports.zoneinfo==0.2.1;python_version<"3.9"
certifi==2022.12.7
charset-normalizer==3.1.0
decorator==5.1.1
Django==4.1.6
django-cors-headers==3.13.0
django-filter==22.1
djangorestframework==3.14.0
executing==1.2.0
Faker==12.0.1
idna==3.4
importlib-metadata==6.0.0
ipython==8.11.0
jedi==0.18.2
Markdown==3.4.1
matplotlib-inline==0.1.6
mixer==7.2.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.38
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.14.0
python-dateutil==2.8.2
pytz==2022.7.1
requests==2.28.2
six==1.16.0
sqlparse==0.4.3
stack-data==0.6.2
traitlets==5.9.0
urllib3==1.26.15
wcwidth==0.2.6
zipp==3.13.0
